BRAND NEW COMMIT! This time with 125% More InvasionMicroGame.
A large chunk of the game logic and physics have been implemented. I
decided to make some GameObject like classes with basic physics. At this
point in time they don't have a super class, but I might get to that
later. I have created a new package for these, which might be temporary,
but it helps me program if I separate things for right now.

A lot of work still needs to be done to the game though.